Correction of a few small bugs

+++ src/engines_gpl/part/part_data/partmem.f90
@@ -385,7 +385,7 @@ module spec_feat_par
       character( 256)          :: leeway_csvfile    ! filename containing predefined leeway factors
       character( 20)           :: leeway_id         ! identifier of the leeway object
       real     ( sp)           :: leeway_multiplier ! leeway multiplier (factor)
-      real     ( sp)           :: leeway_modifier   ! leeway modifier (correction)
-      real     ( sp)           :: leeway_modifier_rad ! leeway modifier (correction) for spherical models
+      real     ( dp)           :: leeway_modifier   ! leeway modifier (correction)
+      real     ( dp)           :: leeway_modifier_rad ! leeway modifier (correction) for spherical models
       real     ( sp)           :: leeway_angle      ! leeway divergence angle
 end module spec_feat_par
